+++ src/runtime/NewReference.cs
@@ -2,6 +2,7 @@ namespace Python.Runtime
 {
     using System;
     using System.Diagnostics.Contracts;
+    using System.Runtime.CompilerServices;
 
     /// <summary>
     /// Represents a reference to a Python object, that is tracked by Python's reference counting.
@@ -56,6 +57,34 @@ namespace Python.Runtime
             return result;
         }
 
+        /// <summary>
+        /// Returns <see cref="PyObject"/> wrapper around this reference, which now owns
+        /// the pointer. Sets the original reference to <c>null</c>, as it no longer owns it.
+        /// </summary>
+        public PyObject MoveToPyObjectOrNull() => this.IsNull() ? null : this.MoveToPyObject();
+        /// <summary>
+        /// Call this method to move ownership of this reference to a Python C API function,
+        /// that steals reference passed to it.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public StolenReference StealNullable()
+        {
+            IntPtr rawPointer = this.pointer;
+            this.pointer = IntPtr.Zero;
+            return new StolenReference(rawPointer);
+        }
+
+        /// <summary>
+        /// Call this method to move ownership of this reference to a Python C API function,
+        /// that steals reference passed to it.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public StolenReference Steal()
+        {
+            if (this.IsNull()) throw new NullReferenceException();
+
+            return this.StealNullable();
+        }
         /// <summary>
         /// Removes this reference to a Python object, and sets it to <c>null</c>.
         /// </summary>