+++ src/runtime/exceptions.cs
@@ -1,7 +1,8 @@
 using System;
+using System.Diagnostics;
 using System.Reflection;
+using System.Runtime.ExceptionServices;
 using System.Runtime.InteropServices;
-using System.Text;
 
 namespace Python.Runtime
 {
@@ -23,19 +24,10 @@ namespace Python.Runtime
         {
         }
 
-        internal static Exception ToException(IntPtr ob)
+        internal static Exception ToException(BorrowedReference ob)
         {
             var co = GetManagedObject(ob) as CLRObject;
-            if (co == null)
-            {
-                return null;
-            }
-            var e = co.inst as Exception;
-            if (e == null)
-            {
-                return null;
-            }
-            return e;
+            return co?.inst as Exception;
         }
 
         /// <summary>
@@ -43,7 +35,7 @@ namespace Python.Runtime
         /// </summary>
         public new static IntPtr tp_repr(IntPtr ob)
         {
-            Exception e = ToException(ob);
+            Exception e = ToException(new BorrowedReference(ob));
             if (e == null)
             {
                 return Exceptions.RaiseTypeError("invalid object");
@@ -66,7 +58,7 @@ namespace Python.Runtime
         /// </summary>
         public new static IntPtr tp_str(IntPtr ob)
         {
-            Exception e = ToException(ob);
+            Exception e = ToException(new BorrowedReference(ob));
             if (e == null)
             {
                 return Exceptions.RaiseTypeError("invalid object");
@@ -93,7 +85,7 @@ namespace Python.Runtime
     /// <remarks>
     /// Readability of the Exceptions class improvements as we look toward version 2.7 ...
     /// </remarks>
-    public static class Exceptions
+    internal static class Exceptions
     {
         internal static PyModule warnings_module;
         internal static PyModule exceptions_module;
@@ -155,8 +147,7 @@ namespace Python.Runtime
         /// __getattr__ implementation, and thus dereferencing a NULL
         /// pointer.
         /// </summary>
-        /// <param name="ob">The python object wrapping </param>
-        internal static void SetArgsAndCause(Exception e, IntPtr ob)
+        internal static void SetArgsAndCause(BorrowedReference ob, Exception e)
         {
             IntPtr args;
             if (!string.IsNullOrEmpty(e.Message))
@@ -170,14 +161,16 @@ namespace Python.Runtime
                 args = Runtime.PyTuple_New(0);
             }
 
-            if (Runtime.PyObject_SetAttrString(ob, "args", args) != 0)
-                throw new PythonException();
+            using var argsTuple = NewReference.DangerousFromPointer(args);
+
+            if (Runtime.PyObject_SetAttrString(ob, "args", argsTuple) != 0)
+                throw PythonException.ThrowLastAsClrException();
 
             if (e.InnerException != null)
             {
                 // Note: For an AggregateException, InnerException is only the first of the InnerExceptions.
                 using var cause = CLRObject.GetReference(e.InnerException);
-                Runtime.PyException_SetCause(ob, cause.DangerousMoveToPointer());
+                Runtime.PyException_SetCause(ob, cause.Steal());
             }
         }
 
@@ -189,7 +182,7 @@ namespace Python.Runtime
         {
             if (pointer.IsNull)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -202,7 +195,7 @@ namespace Python.Runtime
         {
             if (pointer == IntPtr.Zero || ErrorOccurred())
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -210,7 +203,7 @@ namespace Python.Runtime
         {
             if (pointer == IntPtr.Zero && ErrorOccurred())
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return pointer;
         }
@@ -225,19 +218,6 @@ namespace Python.Runtime
         public static bool ExceptionMatches(IntPtr ob)
         {
             return Runtime.PyErr_ExceptionMatches(ob) != 0;
-        }
-
-        /// <summary>
-        /// ExceptionMatches Method
-        /// </summary>
-        /// <remarks>
-        /// Returns true if the given Python exception matches the given
-        /// Python object. This is a wrapper for PyErr_GivenExceptionMatches.
-        /// </remarks>
-        public static bool ExceptionMatches(IntPtr exc, IntPtr ob)
-        {
-            int i = Runtime.PyErr_GivenExceptionMatches(exc, ob);
-            return i != 0;
         }
 
         /// <summary>
@@ -264,6 +244,7 @@ namespace Python.Runtime
             Runtime.PyErr_SetObject(new BorrowedReference(type), new BorrowedReference(exceptionObject));
         }
 
+        internal const string DispatchInfoAttribute = "__dispatch_info__";
         /// <summary>
         /// SetError Method
         /// </summary>
@@ -272,8 +253,10 @@ namespace Python.Runtime
         /// object. The CLR exception instance is wrapped as a Python
         /// object, allowing it to be handled naturally from Python.
         /// </remarks>
-        public static void SetError(Exception e)
+        public static bool SetError(Exception e)
         {
+            Debug.Assert(e is not null);
+
             // Because delegates allow arbitrary nesting of Python calling
             // managed calling Python calling... etc. it is possible that we
             // might get a managed exception raised that is a wrapper for a
@@ -282,31 +265,36 @@ namespace Python.Runtime
             var pe = e as PythonException;
             if (pe != null)
             {
-                Runtime.XIncref(pe.PyType);
-                Runtime.XIncref(pe.PyValue);
-                Runtime.XIncref(pe.PyTB);
-                Runtime.PyErr_Restore(pe.PyType, pe.PyValue, pe.PyTB);
-                return;
+                pe.Restore();
+                return true;
             }
 
-            IntPtr op = CLRObject.GetInstHandle(e);
-            IntPtr etype = Runtime.PyObject_GetAttr(op, PyIdentifier.__class__);
-            Runtime.PyErr_SetObject(new BorrowedReference(etype), new BorrowedReference(op));
-            Runtime.XDecref(etype);
-            Runtime.XDecref(op);
+            using var instance = Converter.ToPythonReference(e);
+            if (instance.IsNull()) return false;
+
+            var exceptionInfo = ExceptionDispatchInfo.Capture(e);
+            using var pyInfo = Converter.ToPythonReference(exceptionInfo);
+
+            if (Runtime.PyObject_SetAttrString(instance, DispatchInfoAttribute, pyInfo) != 0)
+                return false;
+
+            Debug.Assert(Runtime.PyObject_TypeCheck(instance, new BorrowedReference(BaseException)));
+
+            var type = Runtime.PyObject_TYPE(instance);
+            Runtime.PyErr_SetObject(type, instance);
+            return true;
         }
 
         /// <summary>
         /// When called after SetError, sets the cause of the error.
         /// </summary>
         /// <param name="cause">The cause of the current error</param>
-        public static void SetCause(PythonException cause)
+        public static void SetCause(Exception cause)
         {
-            var currentException = new PythonException();
+            var currentException = PythonException.FetchCurrentRaw();
             currentException.Normalize();
-            cause.Normalize();
-            Runtime.XIncref(cause.PyValue);
-            Runtime.PyException_SetCause(currentException.PyValue, cause.PyValue);
+            using var causeInstance = Converter.ToPythonReference(cause);
+            Runtime.PyException_SetCause(currentException.Value!.Reference, causeInstance.Steal());
             currentException.Restore();
         }
 
@@ -319,7 +307,7 @@ namespace Python.Runtime
         /// </remarks>
         public static bool ErrorOccurred()
         {
-            return Runtime.PyErr_Occurred() != IntPtr.Zero;
+            return Runtime.PyErr_Occurred() != null;
         }
 
         /// <summary>
@@ -393,16 +381,21 @@ namespace Python.Runtime
         /// <returns><c>IntPtr.Zero</c></returns>
         internal static IntPtr RaiseTypeError(string message)
         {
-            PythonException previousException = null;
-            if (ErrorOccurred())
-            {
-                previousException = new PythonException();
-            }
+            var cause = PythonException.FetchCurrentOrNullRaw();
+            cause?.Normalize();
+
             Exceptions.SetError(Exceptions.TypeError, message);
-            if (previousException != null)
-            {
-                SetCause(previousException);
-            }
+
+            if (cause is null) return IntPtr.Zero;
+
+            var typeError = PythonException.FetchCurrentRaw();
+            typeError.Normalize();
+
+            Runtime.PyException_SetCause(
+                typeError.Value!.Reference,
+                new NewReference(cause.Value!.Reference).Steal());
+            typeError.Restore();
+
             return IntPtr.Zero;
         }