+++ src/runtime/pylist.cs
@@ -61,7 +61,7 @@ namespace Python.Runtime
         {
             if (obj == IntPtr.Zero)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -77,7 +77,7 @@ namespace Python.Runtime
                 if (r < 0)
                 {
                     Runtime.Py_DecRef(val);
-                    throw new PythonException();
+                    throw PythonException.ThrowLastAsClrException();
                 }
             }
             return val;
@@ -118,7 +118,7 @@ namespace Python.Runtime
             IntPtr op = Runtime.PySequence_List(value.obj);
             if (op == IntPtr.Zero)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return new PyList(op);
         }
@@ -135,7 +135,7 @@ namespace Python.Runtime
             int r = Runtime.PyList_Append(this.Reference, item.obj);
             if (r < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -150,7 +150,7 @@ namespace Python.Runtime
             int r = Runtime.PyList_Insert(this.Reference, index, item.obj);
             if (r < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -166,7 +166,7 @@ namespace Python.Runtime
             int r = Runtime.PyList_Reverse(this.Reference);
             if (r < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -182,7 +182,7 @@ namespace Python.Runtime
             int r = Runtime.PyList_Sort(this.Reference);
             if (r < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
     }