+++ src/runtime/pyobject.cs
@@ -27,7 +27,8 @@ namespace Python.Runtime
 
         protected internal IntPtr obj = IntPtr.Zero;
 
-        internal BorrowedReference Reference => new BorrowedReference(obj);
+        public static PyObject None => new PyObject(new BorrowedReference(Runtime.PyNone));
+        internal BorrowedReference Reference => new BorrowedReference(this.obj);
 
         /// <summary>
         /// PyObject Constructor
@@ -72,6 +73,17 @@ namespace Python.Runtime
             if (reference.IsNull) throw new ArgumentNullException(nameof(reference));
 
             obj = Runtime.SelfIncRef(reference.DangerousGetAddress());
+            Finalizer.Instance.ThrottledCollect();
+#if TRACE_ALLOC
+            Traceback = new StackTrace(1);
+#endif
+        }
+
+        internal PyObject(StolenReference reference)
+        {
+            if (reference == null) throw new ArgumentNullException(nameof(reference));
+
+            obj = reference.DangerousGetAddressOrNull();
             Finalizer.Instance.ThrottledCollect();
 #if TRACE_ALLOC
             Traceback = new StackTrace(1);
@@ -123,6 +135,13 @@ namespace Python.Runtime
             return new PyObject(op);
         }
 
+        /// <summary>
+        /// Creates new <see cref="PyObject"/> from a nullable reference.
+        /// When <paramref name="reference"/> is <c>null</c>, <c>null</c> is returned.
+        /// </summary>
+        internal static PyObject FromNullableReference(BorrowedReference reference)
+            => reference.IsNull ? null : new PyObject(reference);
+
 
         /// <summary>
         /// AsManagedObject Method
@@ -136,7 +155,8 @@ namespace Python.Runtime
             object result;
             if (!Converter.ToManaged(obj, t, out result, true))
             {
-                throw new InvalidCastException("cannot convert object to target type", new PythonException());
+                throw new InvalidCastException("cannot convert object to target type",
+                    PythonException.FetchCurrentOrNull(out _));
             }
             return result;
         }
@@ -198,7 +218,7 @@ namespace Python.Runtime
                     {
                         // Python requires finalizers to preserve exception:
                         // https://docs.python.org/3/extending/newtypes.html#finalization-and-de-allocation
-                        Runtime.PyErr_Restore(errType, errVal, traceback);
+                        Runtime.PyErr_Restore(errType.StealNullable(), errVal.StealNullable(), traceback.StealNullable());
                     }
                 }
                 else
@@ -218,6 +238,9 @@ namespace Python.Runtime
             Dispose(true);
             GC.SuppressFinalize(this);
         }
+
+        internal BorrowedReference GetPythonTypeReference()
+            => new BorrowedReference(Runtime.PyObject_TYPE(obj));
 
         /// <summary>
         /// GetPythonType Method
@@ -291,7 +314,7 @@ namespace Python.Runtime
             IntPtr op = Runtime.PyObject_GetAttrString(obj, name);
             if (op == IntPtr.Zero)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return new PyObject(op);
         }
@@ -333,7 +356,7 @@ namespace Python.Runtime
             IntPtr op = Runtime.PyObject_GetAttr(obj, name.obj);
             if (op == IntPtr.Zero)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return new PyObject(op);
         }
@@ -376,7 +399,7 @@ namespace Python.Runtime
             int r = Runtime.PyObject_SetAttrString(obj, name, value.obj);
             if (r < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -397,7 +420,7 @@ namespace Python.Runtime
             int r = Runtime.PyObject_SetAttr(obj, name.obj, value.obj);
             if (r < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -416,7 +439,7 @@ namespace Python.Runtime
             int r = Runtime.PyObject_SetAttrString(obj, name, IntPtr.Zero);
             if (r < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -436,7 +459,7 @@ namespace Python.Runtime
             int r = Runtime.PyObject_SetAttr(obj, name.obj, IntPtr.Zero);
             if (r < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -456,7 +479,7 @@ namespace Python.Runtime
             IntPtr op = Runtime.PyObject_GetItem(obj, key.obj);
             if (op == IntPtr.Zero)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return new PyObject(op);
         }
@@ -514,7 +537,7 @@ namespace Python.Runtime
             int r = Runtime.PyObject_SetItem(obj, key.obj, value.obj);
             if (r < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -573,7 +596,7 @@ namespace Python.Runtime
             int r = Runtime.PyObject_DelItem(obj, key.obj);
             if (r < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -623,7 +646,7 @@ namespace Python.Runtime
             var s = Runtime.PyObject_Size(Reference);
             if (s < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return s;
         }
@@ -684,7 +707,7 @@ namespace Python.Runtime
             IntPtr r = Runtime.PyObject_GetIter(obj);
             if (r == IntPtr.Zero)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return new PyObject(r);
         }
@@ -721,7 +744,7 @@ namespace Python.Runtime
             t.Dispose();
             if (r == IntPtr.Zero)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return new PyObject(r);
         }
@@ -741,7 +764,7 @@ namespace Python.Runtime
             IntPtr r = Runtime.PyObject_Call(obj, args.obj, IntPtr.Zero);
             if (r == IntPtr.Zero)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return new PyObject(r);
         }
@@ -764,7 +787,7 @@ namespace Python.Runtime
             t.Dispose();
             if (r == IntPtr.Zero)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return new PyObject(r);
         }
@@ -784,7 +807,7 @@ namespace Python.Runtime
             IntPtr r = Runtime.PyObject_Call(obj, args.obj, kw?.obj ?? IntPtr.Zero);
             if (r == IntPtr.Zero)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return new PyObject(r);
         }
@@ -1009,7 +1032,7 @@ namespace Python.Runtime
             IntPtr r = Runtime.PyObject_Dir(obj);
             if (r == IntPtr.Zero)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return new PyList(r);
         }
@@ -1067,7 +1090,7 @@ namespace Python.Runtime
             int r = Runtime.PyObject_Compare(obj, ((PyObject)o).obj);
             if (Exceptions.ErrorOccurred())
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             return r == 0;
         }
@@ -1121,7 +1144,7 @@ namespace Python.Runtime
             int r = Runtime.PyObject_SetAttrString(obj, binder.Name, ptr);
             if (r < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
             Runtime.XDecref(ptr);
             return true;
@@ -1193,7 +1216,7 @@ namespace Python.Runtime
 
             if (Runtime.PyTuple_SetItem(argtuple, i, ptr) < 0)
             {
-                throw new PythonException();
+                throw PythonException.ThrowLastAsClrException();
             }
         }
 
@@ -1440,5 +1463,14 @@ namespace Python.Runtime
                 yield return pyObj.ToString();
             }
         }
+    }
+
+    internal static class PyObjectExtensions
+    {
+        internal static NewReference NewReferenceOrNull(this PyObject self)
+            => NewReference.DangerousFromPointer(
+                (self?.obj ?? IntPtr.Zero) == IntPtr.Zero
+                    ? IntPtr.Zero
+                    : Runtime.SelfIncRef(self.obj));
     }
 }